2.7 Исправлены замечания наставника
