+++ src/view/trip-point.js
@@ -1,15 +1,5 @@
 import dayjs from "dayjs";
-import {addZeroToNumber} from "../utils.js";
-
-const getDateDiff = (start, finish) => {
-  const diffTimeInMs = finish.diff(start);
-  const timeDuration = dayjs.duration(diffTimeInMs);
-  const days = timeDuration.days();
-  const hours = timeDuration.hours();
-  const minutes = timeDuration.minutes();
-  const time = `${(days > 0) ? addZeroToNumber(days) + `D ` : ``}${(hours > 0) ? addZeroToNumber(hours) + `H ` : ``}${(minutes > 0) ? addZeroToNumber(minutes) + `M` : ``}`;
-  return time;
-};
+import {getDateDiff} from "../utils.js";
 
 const generateOffersList = (offersList) => {
   let str = ``;
@@ -25,9 +15,10 @@ const generateOffersList = (offersList) => {
 };
 
 export const createTripPointTemplate = (point) => {
-  const {times, type, destination, offers} = point;
+  const {times, type, destination, offers, isFavorite} = point;
   const {iconSrc, name, price} = type;
   const {start, finish} = times;
+  const favorite = isFavorite ? `event__favorite-btn--active` : ``;
   return `<li class="trip-events__item">
   <div class="event">
     <time class="event__date" datetime="${dayjs(start).format(`YYYY-MM-DD`)}">${dayjs(start).format(`MMM DD`)}</time>
@@ -50,7 +41,7 @@ export const createTripPointTemplate = (point) => {
     <ul class="event__selected-offers">
       ${generateOffersList(offers)}
     </ul>
-    <button class="event__favorite-btn event__favorite-btn--active" type="button">
+    <button class="event__favorite-btn ${favorite}" type="button">
       <span class="visually-hidden">Add to favorite</span>
       <svg class="event__favorite-icon" width="28" height="28" viewBox="0 0 28 28">
         <path d="M14 21l-8.22899 4.3262 1.57159-9.1631L.685209 9.67376 9.8855 8.33688 14 0l4.1145 8.33688 9.2003 1.33688-6.6574 6.48934 1.5716 9.1631L14 21z"/>
